+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3](https://gitlab.com/AquaticNightmare/ez_baker/-/releases/0_0_3)
+### Fixed
+- Memory leak when showing/hiding the cage of a bake group
+- Low to Low baking would ignore the material setup of the low objects
+### Changed
+- The icon for opening the export folder has been changed
+- Lowered the default cage displacement to 0.05 and made the slider more precise
+
 ## [0.0.2](https://gitlab.com/AquaticNightmare/ez_baker/-/releases/0_0_2)
 ### Fixed
 - Missing icon making the UI unusable